Feature/inv/cpanel install (#3)
* cbdi - docs: Updated Readme with project installation instructions

Extended the Readme.md file with more comprehensive instructions for the installation of the project on different platforms (Docker, Cpanel, XAMPP). Additionally, project name was updated from "Project" to "Project Inventory".

* cbdi - feat: Added .htaccess file to inventory directory

A new .htaccess file has been added to the inventory directory to handle URL rewriting. This is particularly useful in hiding the "public" portion of our URLs, making them more user-friendly and clean.

* cbdi - chore: Add zip files

* Add download links and production mode instructions to readme

This commit adds direct download links for `db.zip` and `inventory.zip` in the `readme.md` file. Additionally, it includes instructions to set `APP_DEBUG` to `false` for production mode.

* cbdi - fix: Update download links in readme.md

The commit updates the download links for `db.zip` and `inventory.zip` in the readme.md file from relative to absolute URLs. This change allows for direct downloading of these files enhancing user convenience.

* cbdi - docs: Add PHP version setup instruction in readme.md

A new instruction for setting up the PHP version in cPanel was added to the readme.md file. The specific version recommended is 7.4, specifically the ea-7.4 (EasyApache 4) variant. This update provides a clearer setup procedure for users.